feat: complete validation, fix radius

+++ src/model/node.hpp
@@ -36,7 +36,7 @@ class Node {
 
   /* Methods */
 
-  unsigned radius() const;
+  double radius() const;
 
  private:
   NodeType type;
